+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Centralized utility functions to Core/Utils.lua (truncate, setfontcolor, ...)
 - Label hover behavior: created AttachLabelHighlight utility
 - CreateRadioButtonGroup
+- CreateIconWithLabel: unified icon+label
 
 ## [2.18] - FuBar plugin 1.0